+++ src/libs/hooks/useContextModal.ts
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 
-import { ModalDispatchContext } from '@/libs/contexts/ModalContextProvider';
+import { ModalDispatchContext } from '@/libs/contexts/ModalContext';
 
 export default function useContextModal() {
   const { open: openModal, close: closeModal } = useContext(ModalDispatchContext);
